cr: consolidate calendar creation logic, move to route card context, drop redundent copy func.
