+++ app/i18n/ar.json
@@ -50,7 +50,7 @@
     "routeDetails": "تفاصيل المسار",
     "routeActions": "إجراءات المسار",
     "copyRoute": "نسخ المسار",
-    "share": "مشاركة",
+    "share": "مشاركة تفاصيل الرحلة",
     "routeCopied": "تم نسخ المسار",
     "addedToCalendar": "تمت الإضافة للتقويم"
   },
