+++ app/i18n/en.json
@@ -56,7 +56,7 @@
     "routeDetails": "Route Details",
     "routeActions": "Actions",
     "copyRoute": "Copy",
-    "share": "Share",
+    "share": "Share Ride Details",
     "routeCopied": "Copied to clipboard",
     "addedToCalendar": "Added to calendar"
   },
