+++ app/i18n/he.json
@@ -56,7 +56,7 @@
     "routeDetails": "פרטי מסלול",
     "routeActions": "פעולות",
     "copyRoute": "העתקה",
-    "share": "שתף",
+    "share": "שתף פרטי נסיעה",
     "routeCopied": "המסלול הועתק",
     "addedToCalendar": "נוסף ליומן"
   },
