+++ app/i18n/ru.json
@@ -53,7 +53,7 @@
     "routeDetails": "Детали маршрута",
     "routeActions": "Действия",
     "copyRoute": "Копировать",
-    "share": "Поделиться",
+    "share": "Поделиться деталями поездки",
     "routeCopied": "Скопирован",
     "addedToCalendar": "Добавлено в календарь"
   },
